+++ packages/vm/core/evm/evmtest/evm_test.go
@@ -47,7 +47,7 @@ import (
 )
 
 func TestStorageContract(t *testing.T) {
-	env := initEVM(t)
+	env := InitEVM(t)
 	ethKey, _ := env.soloChain.EthereumAccountByIndexWithL2Funds(0)
 	require.EqualValues(t, 1, env.getBlockNumber()) // evm block number is incremented along with ISC block index
 
@@ -93,7 +93,7 @@ func TestStorageContract(t *testing.T) {
 }
 
 func TestLowLevelCallRevert(t *testing.T) {
-	env := initEVM(t)
+	env := InitEVM(t)
 	ethKey, _ := env.soloChain.NewEthereumAccountWithL2Funds()
 
 	contract := env.deployContract(ethKey, evmtest.RevertTestContractABI, evmtest.RevertTestContractBytecode)
@@ -112,7 +112,7 @@ func TestLowLevelCallRevert(t *testing.T) {
 }
 
 func TestERC20Contract(t *testing.T) {
-	env := initEVM(t)
+	env := InitEVM(t)
 	ethKey, _ := env.soloChain.NewEthereumAccountWithL2Funds()
 
 	// deploy solidity `erc20` contract
@@ -141,7 +141,7 @@ func TestERC20Contract(t *testing.T) {
 }
 
 func TestGetCode(t *testing.T) {
-	env := initEVM(t)
+	env := InitEVM(t)
 	ethKey, _ := env.soloChain.NewEthereumAccountWithL2Funds()
 	erc20 := env.deployERC20Contract(ethKey, "TestCoin", "TEST")
 
@@ -153,7 +153,7 @@ func TestGetCode(t *testing.T) {
 }
 
 func TestGasCharged(t *testing.T) {
-	env := initEVM(t)
+	env := InitEVM(t)
 	ethKey, _ := env.soloChain.NewEthereumAccountWithL2Funds()
 	storage := env.deployStorageContract(ethKey)
 
@@ -169,7 +169,7 @@ func TestGasCharged(t *testing.T) {
 }
 
 func TestGasRatio(t *testing.T) {
-	env := initEVM(t)
+	env := InitEVM(t)
 	ethKey, _ := env.soloChain.NewEthereumAccountWithL2Funds()
 	storage := env.deployStorageContract(ethKey)
 
@@ -199,7 +199,7 @@ func TestGasRatio(t *testing.T) {
 
 // tests that the gas limits are correctly enforced based on the base tokens sent
 func TestGasLimit(t *testing.T) {
-	env := initEVM(t)
+	env := InitEVM(t)
 	ethKey, _ := env.soloChain.NewEthereumAccountWithL2Funds()
 	storage := env.deployStorageContract(ethKey)
 
@@ -223,7 +223,7 @@ func TestGasLimit(t *testing.T) {
 }
 
 func TestNotEnoughISCGas(t *testing.T) {
-	env := initEVM(t)
+	env := InitEVM(t)
 	ethKey, ethAddress := env.soloChain.NewEthereumAccountWithL2Funds()
 	storage := env.deployStorageContract(ethKey)
 
@@ -265,7 +265,7 @@ func TestNotEnoughISCGas(t *testing.T) {
 
 // ensure the amount of base tokens sent impacts the amount of gas used
 func TestLoop(t *testing.T) {
-	env := initEVM(t)
+	env := InitEVM(t)
 	ethKey, _ := env.soloChain.NewEthereumAccountWithL2Funds()
 	loop := env.deployLoopContract(ethKey)
 
@@ -291,7 +291,7 @@ func TestLoop(t *testing.T) {
 }
 
 func TestLoopWithGasLeft(t *testing.T) {
-	env := initEVM(t)
+	env := InitEVM(t)
 	ethKey, _ := env.soloChain.NewEthereumAccountWithL2Funds()
 	iscTest := env.deployISCTestContract(ethKey)
 
@@ -312,7 +312,7 @@ func TestLoopWithGasLeft(t *testing.T) {
 }
 
 func TestEstimateGasWithoutFunds(t *testing.T) {
-	env := initEVM(t)
+	env := InitEVM(t)
 	ethKey, _ := env.soloChain.NewEthereumAccountWithL2Funds()
 	iscTest := env.deployISCTestContract(ethKey)
 
@@ -329,7 +329,7 @@ func TestEstimateGasWithoutFunds(t *testing.T) {
 }
 
 func TestLoopWithGasLeftEstimateGas(t *testing.T) {
-	env := initEVM(t)
+	env := InitEVM(t)
 	ethKey, ethAddr := env.soloChain.NewEthereumAccountWithL2Funds()
 	iscTest := env.deployISCTestContract(ethKey)
 
@@ -356,7 +356,7 @@ func TestLoopWithGasLeftEstimateGas(t *testing.T) {
 }
 
 func TestEstimateContractGas(t *testing.T) {
-	env := initEVM(t)
+	env := InitEVM(t)
 	ethKey, ethAddr := env.soloChain.NewEthereumAccountWithL2Funds()
 	contract := env.deployERC20Contract(ethKey, "TEST", "tst")
 
@@ -381,7 +381,7 @@ func TestEstimateContractGas(t *testing.T) {
 }
 
 func TestCallViewGasLimit(t *testing.T) {
-	env := initEVM(t)
+	env := InitEVM(t)
 	ethKey, _ := env.soloChain.NewEthereumAccountWithL2Funds()
 	loop := env.deployLoopContract(ethKey)
 
@@ -400,7 +400,7 @@ func TestCallViewGasLimit(t *testing.T) {
 func TestMagicContract(t *testing.T) {
 	// deploy the evm contract, which starts an EVM chain and automatically
 	// deploys the isc.sol EVM contract at address 0x10740000...
-	env := initEVM(t)
+	env := InitEVM(t)
 	ethKey, _ := env.soloChain.NewEthereumAccountWithL2Funds()
 
 	// deploy the isc-test.sol EVM contract
@@ -415,7 +415,7 @@ func TestMagicContract(t *testing.T) {
 }
 
 func TestISCChainOwnerID(t *testing.T) {
-	env := initEVM(t)
+	env := InitEVM(t)
 	ethKey, _ := env.soloChain.NewEthereumAccountWithL2Funds()
 
 	var ret struct {
@@ -428,7 +428,7 @@ func TestISCChainOwnerID(t *testing.T) {
 }
 
 func TestISCTimestamp(t *testing.T) {
-	env := initEVM(t)
+	env := InitEVM(t)
 	ethKey, _ := env.soloChain.NewEthereumAccountWithL2Funds()
 
 	var ret int64
@@ -442,7 +442,7 @@ func TestISCTimestamp(t *testing.T) {
 }
 
 func TestISCCallView(t *testing.T) {
-	env := initEVM(t)
+	env := InitEVM(t)
 	ethKey, _ := env.soloChain.NewEthereumAccountWithL2Funds()
 
 	ret := new(iscmagic.ISCDict)
@@ -459,7 +459,7 @@ func TestISCCallView(t *testing.T) {
 }
 
 func TestISCNFTData(t *testing.T) {
-	env := initEVM(t)
+	env := InitEVM(t)
 	ethKey, _ := env.soloChain.NewEthereumAccountWithL2Funds()
 
 	// mint an NFT and send it to the chain
@@ -491,7 +491,7 @@ func TestISCNFTData(t *testing.T) {
 }
 
 func TestISCTriggerEvent(t *testing.T) {
-	env := initEVM(t)
+	env := InitEVM(t)
 	ethKey, _ := env.soloChain.NewEthereumAccountWithL2Funds()
 	iscTest := env.deployISCTestContract(ethKey)
 
@@ -508,7 +508,7 @@ func TestISCTriggerEvent(t *testing.T) {
 }
 
 func TestISCTriggerEventThenFail(t *testing.T) {
-	env := initEVM(t)
+	env := InitEVM(t)
 	ethKey, _ := env.soloChain.NewEthereumAccountWithL2Funds()
 	iscTest := env.deployISCTestContract(ethKey)
 
@@ -523,7 +523,7 @@ func TestISCTriggerEventThenFail(t *testing.T) {
 }
 
 func TestISCEntropy(t *testing.T) {
-	env := initEVM(t)
+	env := InitEVM(t)
 	ethKey, _ := env.soloChain.NewEthereumAccountWithL2Funds()
 	iscTest := env.deployISCTestContract(ethKey)
 
@@ -538,7 +538,7 @@ func TestISCEntropy(t *testing.T) {
 }
 
 func TestISCGetRequestID(t *testing.T) {
-	env := initEVM(t)
+	env := InitEVM(t)
 	ethKey, _ := env.soloChain.NewEthereumAccountWithL2Funds()
 	iscTest := env.deployISCTestContract(ethKey)
 
@@ -556,7 +556,7 @@ func TestISCGetRequestID(t *testing.T) {
 }
 
 func TestReceiptOfFailedTxDoesNotContainEvents(t *testing.T) {
-	env := initEVM(t)
+	env := InitEVM(t)
 	ethKey, _ := env.soloChain.NewEthereumAccountWithL2Funds()
 	iscTest := env.deployISCTestContract(ethKey)
 
@@ -576,7 +576,7 @@ func TestReceiptOfFailedTxDoesNotContainEvents(t *testing.T) {
 }
 
 func TestISCGetSenderAccount(t *testing.T) {
-	env := initEVM(t)
+	env := InitEVM(t)
 	ethKey, _ := env.soloChain.NewEthereumAccountWithL2Funds()
 	iscTest := env.deployISCTestContract(ethKey)
 
@@ -589,7 +589,7 @@ func TestISCGetSenderAccount(t *testing.T) {
 }
 
 func TestSendBaseTokens(t *testing.T) {
-	env := initEVM(t)
+	env := InitEVM(t)
 
 	ethKey, ethAddress := env.soloChain.NewEthereumAccountWithL2Funds()
 	_, receiver := env.solo.NewKeyPair()
@@ -639,7 +639,7 @@ func TestSendBaseTokens(t *testing.T) {
 }
 
 func TestSendBaseTokensAnotherChain(t *testing.T) {
-	env := initEVM(t)
+	env := InitEVM(t)
 
 	ethKey, ethAddress := env.soloChain.NewEthereumAccountWithL2Funds()
 	iscTest := env.deployISCTestContract(ethKey)
@@ -681,7 +681,7 @@ func TestSendBaseTokensAnotherChain(t *testing.T) {
 }
 
 func TestCannotDepleteAccount(t *testing.T) {
-	env := initEVM(t)
+	env := InitEVM(t)
 
 	ethKey, ethAddress := env.soloChain.NewEthereumAccountWithL2Funds()
 	_, receiver := env.solo.NewKeyPair()
@@ -720,7 +720,7 @@ func TestCannotDepleteAccount(t *testing.T) {
 }
 
 func TestSendNFT(t *testing.T) {
-	env := initEVM(t)
+	env := InitEVM(t)
 	ethKey, ethAddr := env.soloChain.NewEthereumAccountWithL2Funds()
 	ethAgentID := isc.NewEthereumAddressAgentID(env.soloChain.ChainID, ethAddr)
 
@@ -764,7 +764,7 @@ func TestSendNFT(t *testing.T) {
 }
 
 func TestERC721NFTs(t *testing.T) {
-	env := initEVM(t)
+	env := InitEVM(t)
 	ethKey, ethAddr := env.soloChain.NewEthereumAccountWithL2Funds()
 	ethAgentID := isc.NewEthereumAddressAgentID(env.soloChain.ChainID, ethAddr)
 
@@ -831,7 +831,7 @@ func TestERC721NFTs(t *testing.T) {
 }
 
 func TestERC721NFTCollection(t *testing.T) {
-	env := initEVM(t)
+	env := InitEVM(t)
 
 	collectionOwner, collectionOwnerAddr := env.solo.NewKeyPairWithFunds()
 	err := env.soloChain.DepositBaseTokensToL2(env.solo.L1BaseTokens(collectionOwnerAddr)/2, collectionOwner)
@@ -970,7 +970,7 @@ func TestERC721NFTCollection(t *testing.T) {
 }
 
 func TestISCCall(t *testing.T) {
-	env := initEVM(t, inccounter.Processor)
+	env := InitEVM(t, inccounter.Processor)
 	ethKey, _ := env.soloChain.NewEthereumAccountWithL2Funds()
 	err := env.soloChain.DeployContract(nil, inccounter.Contract.Name, inccounter.Contract.ProgramHash)
 	require.NoError(t, err)
@@ -989,7 +989,7 @@ func TestISCCall(t *testing.T) {
 }
 
 func TestFibonacciContract(t *testing.T) {
-	env := initEVM(t)
+	env := InitEVM(t)
 	ethKey, _ := env.soloChain.NewEthereumAccountWithL2Funds()
 	fibo := env.deployFibonacciContract(ethKey)
 	require.EqualValues(t, 2, env.getBlockNumber())
@@ -1002,7 +1002,7 @@ func TestFibonacciContract(t *testing.T) {
 }
 
 func TestEVMContractOwnsFundsL2Transfer(t *testing.T) {
-	env := initEVM(t)
+	env := InitEVM(t)
 	ethKey, _ := env.soloChain.NewEthereumAccountWithL2Funds()
 	iscTest := env.deployISCTestContract(ethKey)
 
@@ -1029,7 +1029,7 @@ func TestEVMContractOwnsFundsL2Transfer(t *testing.T) {
 }
 
 func TestISCPanic(t *testing.T) {
-	env := initEVM(t)
+	env := InitEVM(t)
 	ethKey, _ := env.soloChain.NewEthereumAccountWithL2Funds()
 
 	iscTest := env.deployISCTestContract(ethKey)
@@ -1046,7 +1046,7 @@ func TestISCPanic(t *testing.T) {
 }
 
 func TestISCSendWithArgs(t *testing.T) {
-	env := initEVM(t, inccounter.Processor)
+	env := InitEVM(t, inccounter.Processor)
 	err := env.soloChain.DeployContract(nil, inccounter.Contract.Name, inccounter.Contract.ProgramHash)
 	require.NoError(t, err)
 
@@ -1100,7 +1100,7 @@ func TestISCSendWithArgs(t *testing.T) {
 }
 
 func TestERC20BaseTokens(t *testing.T) {
-	env := initEVM(t)
+	env := InitEVM(t)
 	ethKey, ethAddr := env.soloChain.NewEthereumAccountWithL2Funds()
 
 	erc20 := env.ERC20BaseTokens(ethKey)
@@ -1198,7 +1198,7 @@ func TestERC20BaseTokens(t *testing.T) {
 }
 
 func TestERC20NativeTokens(t *testing.T) {
-	env := initEVM(t)
+	env := InitEVM(t)
 
 	const (
 		tokenName         = "ERC20 Native Token Test"
@@ -1252,7 +1252,7 @@ func TestERC20NativeTokens(t *testing.T) {
 }
 
 func TestERC20NativeTokensWithExternalFoundry(t *testing.T) {
-	env := initEVM(t)
+	env := InitEVM(t)
 
 	const (
 		tokenName         = "ERC20 Native Token Test"
@@ -1304,8 +1304,8 @@ func TestERC20NativeTokensWithExternalFoundry(t *testing.T) {
 }
 
 func testERC20NativeTokens(
-	env *soloChainEnv,
-	erc20 *iscContractInstance,
+	env *SoloChainEnv,
+	erc20 *IscContractInstance,
 	nativeTokenID iotago.NativeTokenID,
 	tokenName, tokenTickerSymbol string,
 	tokenDecimals uint8,
@@ -1420,7 +1420,7 @@ func testERC20NativeTokens(
 }
 
 func TestERC20NativeTokensLongName(t *testing.T) {
-	env := initEVM(t)
+	env := InitEVM(t)
 
 	var (
 		tokenName         = strings.Repeat("A", 10_000)
@@ -1443,7 +1443,7 @@ func TestERC20NativeTokensLongName(t *testing.T) {
 
 // test withdrawing ALL EVM balance to a L1 address via the magic contract
 func TestEVMWithdrawAll(t *testing.T) {
-	env := initEVM(t)
+	env := InitEVM(t)
 	ethKey, ethAddress := env.soloChain.NewEthereumAccountWithL2Funds()
 	_, receiver := env.solo.NewKeyPair()
 
@@ -1497,7 +1497,7 @@ func TestEVMWithdrawAll(t *testing.T) {
 }
 
 func TestEVMGasPriceMismatch(t *testing.T) {
-	env := initEVM(t)
+	env := InitEVM(t)
 	ethKey, senderAddress := env.soloChain.NewEthereumAccountWithL2Funds()
 
 	// deploy solidity `storage` contract
@@ -1519,7 +1519,7 @@ func TestEVMGasPriceMismatch(t *testing.T) {
 }
 
 func TestEVMTransferBaseTokens(t *testing.T) {
-	env := initEVM(t)
+	env := InitEVM(t)
 	ethKey, ethAddr := env.soloChain.NewEthereumAccountWithL2Funds()
 	_, someEthereumAddr := solo.NewEthereumAccount()
 	someAgentID := isc.NewEthereumAddressAgentID(env.soloChain.ChainID, someEthereumAddr)
@@ -1556,7 +1556,7 @@ func TestEVMTransferBaseTokens(t *testing.T) {
 }
 
 func TestSolidityTransferBaseTokens(t *testing.T) {
-	env := initEVM(t)
+	env := InitEVM(t)
 	ethKey, _ := env.soloChain.NewEthereumAccountWithL2Funds()
 	_, someEthereumAddr := solo.NewEthereumAccount()
 	someEthereumAgentID := isc.NewEthereumAddressAgentID(env.soloChain.ChainID, someEthereumAddr)
@@ -1639,7 +1639,7 @@ func TestSolidityTransferBaseTokens(t *testing.T) {
 }
 
 func TestSendEntireBalance(t *testing.T) {
-	env := initEVM(t)
+	env := InitEVM(t)
 	ethKey, ethAddr := env.soloChain.NewEthereumAccountWithL2Funds()
 	_, someEthereumAddr := solo.NewEthereumAccount()
 	someEthereumAgentID := isc.NewEthereumAddressAgentID(env.soloChain.ChainID, someEthereumAddr)
@@ -1699,7 +1699,7 @@ func TestSendEntireBalance(t *testing.T) {
 }
 
 func TestSolidityRevertMessage(t *testing.T) {
-	env := initEVM(t)
+	env := InitEVM(t)
 	ethKey, ethAddr := env.soloChain.NewEthereumAccountWithL2Funds()
 	iscTest := env.deployISCTestContract(ethKey)
 
@@ -1729,7 +1729,7 @@ func TestSolidityRevertMessage(t *testing.T) {
 }
 
 func TestCallContractCannotCauseStackOverflow(t *testing.T) {
-	env := initEVM(t)
+	env := InitEVM(t)
 	ethKey, _ := env.soloChain.NewEthereumAccountWithL2Funds()
 
 	iscTest := env.deployISCTestContract(ethKey)
@@ -1749,7 +1749,7 @@ func TestCallContractCannotCauseStackOverflow(t *testing.T) {
 }
 
 func TestStaticCall(t *testing.T) {
-	env := initEVM(t)
+	env := InitEVM(t)
 	ethKey, _ := env.soloChain.NewEthereumAccountWithL2Funds()
 	iscTest := env.deployISCTestContract(ethKey)
 
@@ -1765,7 +1765,7 @@ func TestStaticCall(t *testing.T) {
 }
 
 func TestSelfDestruct(t *testing.T) {
-	env := initEVM(t)
+	env := InitEVM(t)
 	ethKey, _ := env.soloChain.NewEthereumAccountWithL2Funds()
 
 	iscTest := env.deployISCTestContract(ethKey)
@@ -1793,7 +1793,7 @@ func TestSelfDestruct(t *testing.T) {
 }
 
 func TestChangeGasLimit(t *testing.T) {
-	env := initEVM(t)
+	env := InitEVM(t)
 	ethKey, _ := env.soloChain.NewEthereumAccountWithL2Funds()
 	storage := env.deployStorageContract(ethKey)
 
@@ -1818,7 +1818,7 @@ func TestChangeGasLimit(t *testing.T) {
 }
 
 func TestChangeGasPerToken(t *testing.T) {
-	env := initEVM(t)
+	env := InitEVM(t)
 
 	var fee uint64
 	{
@@ -1850,7 +1850,7 @@ func TestChangeGasPerToken(t *testing.T) {
 }
 
 func TestGasPriceIgnoredInEstimateGas(t *testing.T) {
-	env := initEVM(t)
+	env := InitEVM(t)
 
 	var gasLimit []uint64
 
@@ -1880,7 +1880,7 @@ func TestGasPriceIgnoredInEstimateGas(t *testing.T) {
 
 // calling views via eth_call must not cost gas (still has a maximum budget, but simple view calls should pass)
 func TestEVMCallViewGas(t *testing.T) {
-	env := initEVM(t)
+	env := InitEVM(t)
 
 	// issue a view call from an account with no funds
 	ethKey, _ := solo.NewEthereumAccount()
@@ -1893,7 +1893,7 @@ func TestEVMCallViewGas(t *testing.T) {
 }
 
 func TestGasPrice(t *testing.T) {
-	env := initEVM(t)
+	env := InitEVM(t)
 
 	price1 := env.evmChain.GasPrice().Uint64()
 	require.NotZero(t, price1)
@@ -1931,7 +1931,7 @@ func TestGasPrice(t *testing.T) {
 }
 
 func TestTraceTransaction(t *testing.T) {
-	env := initEVM(t)
+	env := InitEVM(t)
 	ethKey, ethAddr := env.soloChain.NewEthereumAccountWithL2Funds()
 
 	traceLatestTx := func() *jsonrpc.CallFrame {
@@ -1966,7 +1966,7 @@ func TestTraceTransaction(t *testing.T) {
 }
 
 func TestMagicContractExamples(t *testing.T) {
-	env := initEVM(t)
+	env := InitEVM(t)
 	ethKey, _ := env.soloChain.NewEthereumAccountWithL2Funds()
 
 	contract := env.deployERC20ExampleContract(ethKey)
@@ -1994,7 +1994,7 @@ func TestMagicContractExamples(t *testing.T) {
 }
 
 func TestCaller(t *testing.T) {
-	env := initEVM(t)
+	env := InitEVM(t)
 	ethKey, _ := env.soloChain.NewEthereumAccountWithL2Funds()
 	iscTest := env.deployISCTestContract(ethKey)
 	err := env.soloChain.TransferAllowanceTo(
@@ -2013,7 +2013,7 @@ func TestCaller(t *testing.T) {
 }
 
 func TestCustomError(t *testing.T) {
-	env := initEVM(t)
+	env := InitEVM(t)
 	ethKey, _ := env.soloChain.NewEthereumAccountWithL2Funds()
 	iscTest := env.deployISCTestContract(ethKey)
 	_, err := iscTest.callFn([]ethCallOptions{{
@@ -2032,7 +2032,7 @@ func TestCustomError(t *testing.T) {
 }
 
 func TestEmitEventAndRevert(t *testing.T) {
-	env := initEVM(t)
+	env := InitEVM(t)
 	ethKey, _ := env.soloChain.NewEthereumAccountWithL2Funds()
 	iscTest := env.deployISCTestContract(ethKey)
 	res, err := iscTest.callFn([]ethCallOptions{{
