+++ packages/vm/core/evm/evmtest/setup.go
@@ -1,0 +1,370 @@
+package evmtest
+
+import (
+	"crypto/ecdsa"
+	"github.com/ethereum/go-ethereum"
+	"github.com/ethereum/go-ethereum/accounts/abi"
+	"github.com/ethereum/go-ethereum/common"
+	"github.com/ethereum/go-ethereum/core/types"
+	"github.com/ethereum/go-ethereum/crypto"
+	iotago "github.com/iotaledger/iota.go/v3"
+	"github.com/iotaledger/wasp/packages/cryptolib"
+	"github.com/iotaledger/wasp/packages/evm/evmtest"
+	"github.com/iotaledger/wasp/packages/evm/evmutil"
+	"github.com/iotaledger/wasp/packages/evm/jsonrpc"
+	"github.com/iotaledger/wasp/packages/isc"
+	"github.com/iotaledger/wasp/packages/isc/coreutil"
+	"github.com/iotaledger/wasp/packages/kv/codec"
+	"github.com/iotaledger/wasp/packages/kv/dict"
+	"github.com/iotaledger/wasp/packages/solo"
+	"github.com/iotaledger/wasp/packages/util"
+	"github.com/iotaledger/wasp/packages/vm/core/evm"
+	"github.com/iotaledger/wasp/packages/vm/core/evm/iscmagic"
+	"github.com/iotaledger/wasp/packages/vm/core/governance"
+	"github.com/iotaledger/wasp/packages/vm/gas"
+	"github.com/stretchr/testify/require"
+	"math/big"
+	"strings"
+	"testing"
+)
+
+type SoloChainEnv struct {
+	t          testing.TB
+	solo       *solo.Solo
+	soloChain  *solo.Chain
+	evmChainID uint16
+	evmChain   *jsonrpc.EVMChain
+}
+
+type iscCallOptions struct {
+	wallet *cryptolib.KeyPair
+}
+
+type ethCallOptions struct {
+	sender   *ecdsa.PrivateKey
+	value    *big.Int
+	gasLimit uint64
+	gasPrice *big.Int
+}
+
+func InitEVM(t testing.TB, nativeContracts ...*coreutil.ContractProcessor) *SoloChainEnv {
+	env := solo.New(t, &solo.InitOptions{
+		AutoAdjustStorageDeposit: true,
+		Debug:                    true,
+		PrintStackTrace:          true,
+		GasBurnLogEnabled:        false,
+	})
+	for _, c := range nativeContracts {
+		env = env.WithNativeContract(c)
+	}
+	return InitEVMWithSolo(t, env)
+}
+
+func InitEVMWithSolo(t testing.TB, env *solo.Solo) *SoloChainEnv {
+	soloChain, _ := env.NewChainExt(nil, 0, "evmchain")
+	return &SoloChainEnv{
+		t:          t,
+		solo:       env,
+		soloChain:  soloChain,
+		evmChainID: evm.DefaultChainID,
+		evmChain:   soloChain.EVM(),
+	}
+}
+
+func (e *SoloChainEnv) parseISCCallOptions(opts []iscCallOptions) iscCallOptions {
+	if len(opts) == 0 {
+		opts = []iscCallOptions{{}}
+	}
+	opt := opts[0]
+	if opt.wallet == nil {
+		opt.wallet = e.soloChain.OriginatorPrivateKey
+	}
+	return opt
+}
+
+func (e *SoloChainEnv) resolveError(err error) error {
+	if err == nil {
+		return nil
+	}
+	if vmError, ok := err.(*isc.UnresolvedVMError); ok {
+		resolvedErr := e.soloChain.ResolveVMError(vmError)
+		return resolvedErr.AsGoError()
+	}
+	return err
+}
+
+func (e *SoloChainEnv) getBlockNumber() uint64 {
+	n := e.evmChain.BlockNumber()
+	return n.Uint64()
+}
+
+func (e *SoloChainEnv) getCode(addr common.Address) []byte {
+	ret, err := e.evmChain.Code(addr, nil)
+	require.NoError(e.t, err)
+	return ret
+}
+
+func (e *SoloChainEnv) getEVMGasRatio() util.Ratio32 {
+	ret, err := e.soloChain.CallView(governance.Contract.Name, governance.ViewGetEVMGasRatio.Name)
+	require.NoError(e.t, err)
+	ratio, err := codec.DecodeRatio32(ret.Get(governance.ParamEVMGasRatio))
+	require.NoError(e.t, err)
+	return ratio
+}
+
+func (e *SoloChainEnv) setEVMGasRatio(newGasRatio util.Ratio32, opts ...iscCallOptions) error {
+	opt := e.parseISCCallOptions(opts)
+	req := solo.NewCallParams(governance.Contract.Name, governance.FuncSetEVMGasRatio.Name, governance.ParamEVMGasRatio, newGasRatio.Bytes())
+	_, err := e.soloChain.PostRequestSync(req, opt.wallet)
+	return err
+}
+
+func (e *SoloChainEnv) setFeePolicy(p gas.FeePolicy, opts ...iscCallOptions) error { //nolint:unparam
+	opt := e.parseISCCallOptions(opts)
+	req := solo.NewCallParams(
+		governance.Contract.Name, governance.FuncSetFeePolicy.Name,
+		governance.ParamFeePolicyBytes,
+		p.Bytes(),
+	)
+	_, err := e.soloChain.PostRequestSync(req, opt.wallet)
+	return err
+}
+
+func (e *SoloChainEnv) getNonce(addr common.Address) uint64 {
+	nonce, err := e.evmChain.TransactionCount(addr, nil)
+	require.NoError(e.t, err)
+	return nonce
+}
+
+func (e *SoloChainEnv) contractFromABI(address common.Address, abiJSON string, defaultSender *ecdsa.PrivateKey) *IscContractInstance {
+	parsedABI, err := abi.JSON(strings.NewReader(abiJSON))
+	require.NoError(e.t, err)
+	return &IscContractInstance{
+		evmContractInstance: &evmContractInstance{
+			chain:         e,
+			defaultSender: defaultSender,
+			address:       address,
+			abi:           parsedABI,
+		},
+	}
+}
+
+func (e *SoloChainEnv) ISCMagicSandbox(defaultSender *ecdsa.PrivateKey) *IscContractInstance {
+	return e.contractFromABI(iscmagic.Address, iscmagic.SandboxABI, defaultSender)
+}
+
+func (e *SoloChainEnv) ISCMagicUtil(defaultSender *ecdsa.PrivateKey) *IscContractInstance {
+	return e.contractFromABI(iscmagic.Address, iscmagic.UtilABI, defaultSender)
+}
+
+func (e *SoloChainEnv) ISCMagicAccounts(defaultSender *ecdsa.PrivateKey) *IscContractInstance {
+	return e.contractFromABI(iscmagic.Address, iscmagic.AccountsABI, defaultSender)
+}
+
+func (e *SoloChainEnv) ISCMagicPrivileged(defaultSender *ecdsa.PrivateKey) *IscContractInstance {
+	return e.contractFromABI(iscmagic.Address, iscmagic.PrivilegedABI, defaultSender)
+}
+
+func (e *SoloChainEnv) ERC20BaseTokens(defaultSender *ecdsa.PrivateKey) *IscContractInstance {
+	erc20BaseABI, err := abi.JSON(strings.NewReader(iscmagic.ERC20BaseTokensABI))
+	require.NoError(e.t, err)
+	return &IscContractInstance{
+		evmContractInstance: &evmContractInstance{
+			chain:         e,
+			defaultSender: defaultSender,
+			address:       iscmagic.ERC20BaseTokensAddress,
+			abi:           erc20BaseABI,
+		},
+	}
+}
+
+func (e *SoloChainEnv) ERC20NativeTokens(defaultSender *ecdsa.PrivateKey, foundrySN uint32) *IscContractInstance {
+	erc20BaseABI, err := abi.JSON(strings.NewReader(iscmagic.ERC20NativeTokensABI))
+	require.NoError(e.t, err)
+	return &IscContractInstance{
+		evmContractInstance: &evmContractInstance{
+			chain:         e,
+			defaultSender: defaultSender,
+			address:       iscmagic.ERC20NativeTokensAddress(foundrySN),
+			abi:           erc20BaseABI,
+		},
+	}
+}
+
+func (e *SoloChainEnv) ERC20ExternalNativeTokens(defaultSender *ecdsa.PrivateKey, addr common.Address) *IscContractInstance {
+	erc20BaseABI, err := abi.JSON(strings.NewReader(iscmagic.ERC20ExternalNativeTokensABI))
+	require.NoError(e.t, err)
+	return &IscContractInstance{
+		evmContractInstance: &evmContractInstance{
+			chain:         e,
+			defaultSender: defaultSender,
+			address:       addr,
+			abi:           erc20BaseABI,
+		},
+	}
+}
+
+func (e *SoloChainEnv) ERC721NFTs(defaultSender *ecdsa.PrivateKey) *IscContractInstance {
+	erc721ABI, err := abi.JSON(strings.NewReader(iscmagic.ERC721NFTsABI))
+	require.NoError(e.t, err)
+	return &IscContractInstance{
+		evmContractInstance: &evmContractInstance{
+			chain:         e,
+			defaultSender: defaultSender,
+			address:       iscmagic.ERC721NFTsAddress,
+			abi:           erc721ABI,
+		},
+	}
+}
+
+func (e *SoloChainEnv) ERC721NFTCollection(defaultSender *ecdsa.PrivateKey, collectionID iotago.NFTID) *IscContractInstance {
+	erc721NFTCollectionABI, err := abi.JSON(strings.NewReader(iscmagic.ERC721NFTCollectionABI))
+	require.NoError(e.t, err)
+	return &IscContractInstance{
+		evmContractInstance: &evmContractInstance{
+			chain:         e,
+			defaultSender: defaultSender,
+			address:       iscmagic.ERC721NFTCollectionAddress(collectionID),
+			abi:           erc721NFTCollectionABI,
+		},
+	}
+}
+
+func (e *SoloChainEnv) deployISCTestContract(creator *ecdsa.PrivateKey) *iscTestContractInstance {
+	return &iscTestContractInstance{e.deployContract(creator, evmtest.ISCTestContractABI, evmtest.ISCTestContractBytecode)}
+}
+
+func (e *SoloChainEnv) deployStorageContract(creator *ecdsa.PrivateKey) *storageContractInstance {
+	return &storageContractInstance{e.deployContract(creator, evmtest.StorageContractABI, evmtest.StorageContractBytecode, uint32(42))}
+}
+
+func (e *SoloChainEnv) deployERC20Contract(creator *ecdsa.PrivateKey, name, symbol string) *erc20ContractInstance {
+	return &erc20ContractInstance{e.deployContract(creator, evmtest.ERC20ContractABI, evmtest.ERC20ContractBytecode, name, symbol)}
+}
+
+func (e *SoloChainEnv) deployLoopContract(creator *ecdsa.PrivateKey) *loopContractInstance {
+	return &loopContractInstance{e.deployContract(creator, evmtest.LoopContractABI, evmtest.LoopContractBytecode)}
+}
+
+func (e *SoloChainEnv) deployFibonacciContract(creator *ecdsa.PrivateKey) *fibonacciContractInstance {
+	return &fibonacciContractInstance{e.deployContract(creator, evmtest.FibonacciContractABI, evmtest.FibonacciContractByteCode)}
+}
+
+func (e *SoloChainEnv) deployERC20ExampleContract(creator *ecdsa.PrivateKey) *erc20ContractInstance {
+	return &erc20ContractInstance{e.deployContract(creator, evmtest.ERC20ExampleContractABI, evmtest.ERC20ExampleContractBytecode)}
+}
+
+func (e *SoloChainEnv) signer() types.Signer {
+	return evmutil.Signer(big.NewInt(int64(e.evmChainID)))
+}
+
+func (e *SoloChainEnv) maxGasLimit() uint64 {
+	fp := e.soloChain.GetGasFeePolicy()
+	gl := e.soloChain.GetGasLimits()
+	return gas.EVMCallGasLimit(gl, &fp.EVMGasRatio)
+}
+
+func (e *SoloChainEnv) deployContract(creator *ecdsa.PrivateKey, abiJSON string, bytecode []byte, args ...interface{}) *evmContractInstance {
+	creatorAddress := crypto.PubkeyToAddress(creator.PublicKey)
+
+	nonce := e.getNonce(creatorAddress)
+
+	contractABI, err := abi.JSON(strings.NewReader(abiJSON))
+	require.NoError(e.t, err)
+	constructorArguments, err := contractABI.Pack("", args...)
+	require.NoError(e.t, err)
+
+	data := []byte{}
+	data = append(data, bytecode...)
+	data = append(data, constructorArguments...)
+
+	value := big.NewInt(0)
+
+	gasLimit, err := e.evmChain.EstimateGas(ethereum.CallMsg{
+		From:  creatorAddress,
+		Value: value,
+		Data:  data,
+	}, nil)
+	require.NoError(e.t, err)
+
+	tx, err := types.SignTx(
+		types.NewContractCreation(nonce, value, gasLimit, e.evmChain.GasPrice(), data),
+		e.signer(),
+		creator,
+	)
+	require.NoError(e.t, err)
+
+	err = e.evmChain.SendTransaction(tx)
+	require.NoError(e.t, err)
+
+	return &evmContractInstance{
+		chain:         e,
+		defaultSender: creator,
+		address:       crypto.CreateAddress(creatorAddress, nonce),
+		abi:           contractABI,
+	}
+}
+
+func (e *SoloChainEnv) registerERC20NativeToken(
+	foundryOwner *cryptolib.KeyPair,
+	foundrySN uint32,
+	tokenName, tokenTickerSymbol string,
+	tokenDecimals uint8,
+) error {
+	_, err := e.soloChain.PostRequestOffLedger(solo.NewCallParams(evm.Contract.Name, evm.FuncRegisterERC20NativeToken.Name, dict.Dict{
+		evm.FieldFoundrySN:         codec.EncodeUint32(foundrySN),
+		evm.FieldTokenName:         codec.EncodeString(tokenName),
+		evm.FieldTokenTickerSymbol: codec.EncodeString(tokenTickerSymbol),
+		evm.FieldTokenDecimals:     codec.EncodeUint8(tokenDecimals),
+	}).WithMaxAffordableGasBudget(), foundryOwner)
+	return err
+}
+
+func (e *SoloChainEnv) registerERC20ExternalNativeToken(
+	fromChain *solo.Chain,
+	foundrySN uint32,
+	tokenName, tokenTickerSymbol string,
+	tokenDecimals uint8,
+) (ret common.Address, err error) {
+	_, err = fromChain.PostRequestOffLedger(solo.NewCallParams(evm.Contract.Name, evm.FuncRegisterERC20NativeTokenOnRemoteChain.Name, dict.Dict{
+		evm.FieldFoundrySN:         codec.EncodeUint32(foundrySN),
+		evm.FieldTokenName:         codec.EncodeString(tokenName),
+		evm.FieldTokenTickerSymbol: codec.EncodeString(tokenTickerSymbol),
+		evm.FieldTokenDecimals:     codec.EncodeUint8(tokenDecimals),
+		evm.FieldTargetAddress:     codec.EncodeAddress(e.soloChain.ChainID.AsAddress()),
+	}).
+		// to cover sd and gas fee for the 'FuncRegisterERC20ExternalNativeToken' func call in 'FuncRegisterERC20NativeTokenOnRemoteChain'
+		WithAllowance(isc.NewAssetsBaseTokens(20*gas.LimitsDefault.MinGasPerRequest)).
+		WithGasBudget(10*gas.LimitsDefault.MinGasPerRequest),
+		fromChain.OriginatorPrivateKey)
+	if err != nil {
+		return ret, err
+	}
+
+	foundryOutput, err := fromChain.GetFoundryOutput(foundrySN)
+	require.NoError(e.t, err)
+	nativeTokenID, err := foundryOutput.ID()
+	require.NoError(e.t, err)
+
+	if !e.soloChain.WaitUntil(func() bool {
+		res, err2 := e.soloChain.CallView(evm.Contract.Name, evm.FuncGetERC20ExternalNativeTokenAddress.Name,
+			evm.FieldNativeTokenID, nativeTokenID[:],
+		)
+		require.NoError(e.t, err2)
+		if len(res[evm.FieldResult]) == 0 {
+			return false
+		}
+		copy(ret[:], res[evm.FieldResult])
+		return true
+	}) {
+		require.FailNow(e.t, "could not get ERC20 address on target chain")
+	}
+	return ret, err
+}
+
+func (e *SoloChainEnv) registerERC721NFTCollection(collectionOwner *cryptolib.KeyPair, collectionID iotago.NFTID) error {
+	_, err := e.soloChain.PostRequestOffLedger(solo.NewCallParams(evm.Contract.Name, evm.FuncRegisterERC721NFTCollection.Name, dict.Dict{
+		evm.FieldNFTCollectionID: codec.EncodeNFTID(collectionID),
+	}).WithMaxAffordableGasBudget(), collectionOwner)
+	return err
+}
